+++ src/Test/UnitTests/modifications.jl
@@ -1,12 +1,12 @@
 const modificationtests = Dict{String,Function}()
 
 """
-    set_function_single_variable(model::MOI.ModelLike, config::TestConfig)
+    set_function_single_variable(model::MOI.ModelLike, config::Config)
 
 Test that modifying the function of a `SingleVariable`-in-`LessThan` constraint
 throws a [`SettingSingleVariableFunctionNotAllowed`](@ref) error.
 """
-function set_function_single_variable(model::MOI.ModelLike, config::TestConfig)
+function set_function_single_variable(model::MOI.ModelLike, config::Config)
     MOI.empty!(model)
     MOIU.loadfromstring!(
         model,
@@ -29,7 +29,7 @@ end
 modificationtests["set_function_single_variable"] = set_function_single_variable
 
 """
-    solve_set_singlevariable_lessthan(model::MOI.ModelLike, config::TestConfig)
+    solve_set_singlevariable_lessthan(model::MOI.ModelLike, config::Config)
 
 Test set modification SingleVariable-in-LessThan constraint. If
 `config.solve=true` confirm that it solves correctly, and if
@@ -37,7 +37,7 @@ Test set modification SingleVariable-in-LessThan constraint. If
 """
 function solve_set_singlevariable_lessthan(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -74,7 +74,7 @@ modificationtests["solve_set_singlevariable_lessthan"] =
     solve_set_singlevariable_lessthan
 
 """
-    solve_transform_singlevariable_lessthan(model::MOI.ModelLike, config::TestConfig)
+    solve_transform_singlevariable_lessthan(model::MOI.ModelLike, config::Config)
 
 Test set transformation of a SingleVariable-in-LessThan constraint. If
 `config.solve=true` confirm that it solves correctly, and if
@@ -82,7 +82,7 @@ Test set transformation of a SingleVariable-in-LessThan constraint. If
 """
 function solve_transform_singlevariable_lessthan(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -121,7 +121,7 @@ modificationtests["solve_transform_singlevariable_lessthan"] =
     solve_transform_singlevariable_lessthan
 
 """
-    solve_set_scalaraffine_lessthan(model::MOI.ModelLike, config::TestConfig)
+    solve_set_scalaraffine_lessthan(model::MOI.ModelLike, config::Config)
 
 Test modifying set of ScalarAffineFunction-in-LessThan constraint. If
 `config.solve=true` confirm that it solves correctly, and if
@@ -129,7 +129,7 @@ Test modifying set of ScalarAffineFunction-in-LessThan constraint. If
 """
 function solve_set_scalaraffine_lessthan(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -165,7 +165,7 @@ modificationtests["solve_set_scalaraffine_lessthan"] =
     solve_set_scalaraffine_lessthan
 
 """
-    solve_coef_scalaraffine_lessthan(model::MOI.ModelLike, config::TestConfig)
+    solve_coef_scalaraffine_lessthan(model::MOI.ModelLike, config::Config)
 
 Test modifying a variable coefficient in a ScalarAffineFunction-in-LessThan
 constraint. If `config.solve=true` confirm that it solves correctly, and if
@@ -173,7 +173,7 @@ constraint. If `config.solve=true` confirm that it solves correctly, and if
 """
 function solve_coef_scalaraffine_lessthan(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -208,7 +208,7 @@ modificationtests["solve_coef_scalaraffine_lessthan"] =
     solve_coef_scalaraffine_lessthan
 
 """
-    solve_func_scalaraffine_lessthan(model::MOI.ModelLike, config::TestConfig)
+    solve_func_scalaraffine_lessthan(model::MOI.ModelLike, config::Config)
 
 Test setting the function in a ScalarAffineFunction-in-LessThan
 constraint. If `config.solve=true` confirm that it solves correctly, and if
@@ -216,7 +216,7 @@ constraint. If `config.solve=true` confirm that it solves correctly, and if
 """
 function solve_func_scalaraffine_lessthan(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -258,7 +258,7 @@ modificationtests["solve_func_scalaraffine_lessthan"] =
     solve_func_scalaraffine_lessthan
 
 """
-    solve_func_vectoraffine_nonneg(model::MOI.ModelLike, config::TestConfig)
+    solve_func_vectoraffine_nonneg(model::MOI.ModelLike, config::Config)
 
 Test setting the function in a VectorAffineFunction-in-Nonnegatives
 constraint. If `config.solve=true` confirm that it solves correctly, and if
@@ -266,7 +266,7 @@ constraint. If `config.solve=true` confirm that it solves correctly, and if
 """
 function solve_func_vectoraffine_nonneg(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -319,7 +319,7 @@ modificationtests["solve_func_vectoraffine_nonneg"] =
     solve_func_vectoraffine_nonneg
 
 """
-    solve_const_vectoraffine_nonpos(model::MOI.ModelLike, config::TestConfig)
+    solve_const_vectoraffine_nonpos(model::MOI.ModelLike, config::Config)
 
 Test modifying the constant term in a VectorAffineFunction-in-Nonpositives
 constraint. If `config.solve=true` confirm that it solves correctly, and if
@@ -327,7 +327,7 @@ constraint. If `config.solve=true` confirm that it solves correctly, and if
 """
 function solve_const_vectoraffine_nonpos(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -370,7 +370,7 @@ modificationtests["solve_const_vectoraffine_nonpos"] =
     solve_const_vectoraffine_nonpos
 
 """
-    solve_multirow_vectoraffine_nonpos(model::MOI.ModelLike, config::TestConfig)
+    solve_multirow_vectoraffine_nonpos(model::MOI.ModelLike, config::Config)
 
 Test modifying the variable coefficients in a
 VectorAffineFunction-in-Nonpositives constraint. If `config.solve=true` confirm
@@ -378,7 +378,7 @@ that it solves correctly.
 """
 function solve_multirow_vectoraffine_nonpos(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     MOI.empty!(model)
     MOIU.loadfromstring!(
@@ -420,12 +420,12 @@ modificationtests["solve_multirow_vectoraffine_nonpos"] =
     solve_multirow_vectoraffine_nonpos
 
 """
-    solve_const_scalar_objective(model::MOI.ModelLike, config::TestConfig)
+    solve_const_scalar_objective(model::MOI.ModelLike, config::Config)
 
 Test the constant of a scalaraffine objective. If `config.solve=true` confirm
 that it solves correctly.
 """
-function solve_const_scalar_objective(model::MOI.ModelLike, config::TestConfig)
+function solve_const_scalar_objective(model::MOI.ModelLike, config::Config)
     MOI.empty!(model)
     MOIU.loadfromstring!(
         model,
@@ -457,12 +457,12 @@ end
 modificationtests["solve_const_scalar_objective"] = solve_const_scalar_objective
 
 """
-    solve_coef_scalar_objective(model::MOI.ModelLike, config::TestConfig)
+    solve_coef_scalar_objective(model::MOI.ModelLike, config::Config)
 
 Test modifying a variable coefficient in a scalaraffine objective. If
 `config.solve=true` confirm that it solves correctly.
 """
-function solve_coef_scalar_objective(model::MOI.ModelLike, config::TestConfig)
+function solve_coef_scalar_objective(model::MOI.ModelLike, config::Config)
     MOI.empty!(model)
     MOIU.loadfromstring!(
         model,
@@ -495,7 +495,7 @@ modificationtests["solve_coef_scalar_objective"] = solve_coef_scalar_objective
 
 function delete_variable_with_single_variable_obj(
     model::MOI.ModelLike,
-    config::TestConfig,
+    config::Config,
 )
     atol, rtol = config.atol, config.rtol
     MOI.empty!(model)
@@ -528,13 +528,13 @@ modificationtests["delete_variable_with_single_variable_obj"] =
     delete_variable_with_single_variable_obj
 
 """
-    delete_variables_in_a_batch(model::MOI.ModelLike, config::TestConfig)
+    delete_variables_in_a_batch(model::MOI.ModelLike, config::Config)
 
 Test deleting many variables in a batch (i.e. using the delete method which
 takes a vector of variable references). If `config.solve=true` confirm that it
 solves correctly.
 """
-function delete_variables_in_a_batch(model::MOI.ModelLike, config::TestConfig)
+function delete_variables_in_a_batch(model::MOI.ModelLike, config::Config)
     atol = config.atol
     rtol = config.rtol
     MOI.empty!(model)