+++ test/Bridges/Constraint/rsoc.jl
@@ -9,7 +9,7 @@ const MOIB = MathOptInterface.Bridges
 include("../utilities.jl")
 
 mock = MOIU.MockOptimizer(MOIU.UniversalFallback(MOIU.Model{Float64}()))
-config = MOIT.TestConfig()
+config = MOIT.Config()
 
 @testset "RSOC" begin
     bridged_mock = MOIB.Constraint.RSOC{Float64}(mock)